tests/wpt/web-platform-tests/webxr/navigator_xr_requestDevice.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="resources/webxr_util.js"></script>

  <script>
    xr_promise_test("navigator.xr.requestDevice returns a device", (t) => {
      return XRTest.simulateDeviceConnection({ supportsImmersive: true })
        .then( (controller) => { return navigator.xr.requestDevice() })
        .then( (device) => {
          t.step(() => {
            assert_true(device != null);
            assert_true(device instanceof XRDevice);
          });
        });
    });
  </script>
</body>
